moved the other svgs and the forms
